+++ src/read_csv.c
@@ -408,7 +408,7 @@ load_row_values (GList * rows, GGobiData * d, gboolean row_labels)
       gchar *str = row->src->str + row->entry[j + offset].ofs;
       //fprintf(stderr, "string: %s\n", str);
       if (is_numeric (str, row->entry[j + offset].len) && !vt->nlevels)
-        ggobi_data_set_raw_value(d, i, j, (gfloat) g_strtod (str, NULL));
+        ggobi_data_set_raw_value(d, i, j, (gdouble) g_strtod (str, NULL));
       else {
         if (str[0] == '\0' || !g_ascii_strcasecmp (str, "na")
             || !strcmp (str, ".")) {
